Some changes from today and yesterday. Moving towards obstacle avoidance using the light sensors. Detecting obstacles based on if the current light sensor value is below the average light sensor value by a certain amount. Using a task to switch between going straight, and doing an avoidance maneuver. Switched the PID controller to use 16 bit integers since memory was becoming an issue and I wasn't even really using doubles anyways.

+++ lab01/wquade-lab01/PID.ino
@@ -2,23 +2,23 @@
 
 // Function to calculate PID output given set point, current value (process variable), and PID controller parameters.
 // Based on this: https://gist.github.com/bradley219/5373998
-double CalculatePID(double setPoint, double processVariable, struct PID *pid) {
+int8_t CalculatePID(int16_t setPoint, int16_t processVariable, struct PID *pid) {
   // Error
-  double error = setPoint - processVariable;
+  int16_t error = setPoint - processVariable;
 
   // Proportional
-  double proportionalOutput = pid->kp * error;
+  int16_t proportionalOutput = pid->kp * error;
 
   // Integral
   pid->integral += error * pid->dt;
-  double integralOutput = pid->ki * pid->integral;
+  int16_t integralOutput = pid->ki * pid->integral;
 
   // Derivative
-  double derivative = (error - pid->error) / pid->dt;
-  double derivativeOutput = pid->kd * derivative;
+  int16_t derivative = (error - pid->error) / pid->dt;
+  int16_t derivativeOutput = pid->kd * derivative;
 
   // Output
-  double output = proportionalOutput + integralOutput + derivativeOutput;  
+  int16_t output = proportionalOutput + integralOutput + derivativeOutput;  
 
   // Limit to minimum and maximum
   if(output > pid->maximum) {
@@ -30,6 +30,6 @@ double CalculatePID(double setPoint, double processVariable, struct PID *pid) {
   // Save error for derivative
   pid->error = error;
 
-  return output;
+  return (int8_t)output;
 }
